requirements.txt
asgiref==3.4.1
certifi==2024.7.4
distlib==0.3.8
Django==3.2.25
django-ckeditor==6.3.2
django-js-asset==2.0.0
django-markdownx==4.0.7
filelock==3.4.1
importlib-metadata==4.8.3
importlib-resources==5.4.0
Markdown==3.3.7
Pillow==8.4.0
pipenv==2022.4.20
platformdirs==2.4.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.1.1
virtualenv==20.17.1
virtualenv-clone==0.5.7
zipp==3.6.0
